+++ wpull/database/sqltable_test.py
@@ -3,7 +3,7 @@
 
 import unittest
 
-from wpull.database import SQLiteURLTable
+from wpull.database.sqltable import SQLiteURLTable
 from wpull.item import Status
 
 
@@ -18,18 +18,18 @@ class TestDatabase(unittest.TestCase):
             'http://example.com/kitteh',
             'http://example.com/doge',
         ]
-        url_table.add(
+        url_table.add_many(
             urls, referrer='http://example.com', level=0,
             top_url='http://example.net',
         )
 
-        self.assertIn(urls[0], url_table)
-        self.assertIn(urls[1], url_table)
-        self.assertIn(urls[2], url_table)
-        self.assertEqual(3, len(url_table))
+        self.assertTrue(url_table.contains(urls[0]))
+        self.assertTrue(url_table.contains(urls[1]))
+        self.assertTrue(url_table.contains(urls[2]))
+        self.assertEqual(3, url_table.count())
 
         for i in range(3):
-            url_record = url_table[urls[i]]
+            url_record = url_table.get_one(urls[i])
 
             self.assertEqual(urls[i], url_record.url)
             self.assertEqual(Status.todo, url_record.status)
@@ -37,17 +37,16 @@ class TestDatabase(unittest.TestCase):
             self.assertEqual('http://example.com', url_record.referrer)
             self.assertEqual('http://example.net', url_record.top_url)
 
-        url_record = url_table.get_and_update(
+        url_record = url_table.check_out(
             Status.todo,
-            new_status=Status.in_progress
         )
 
         self.assertEqual(Status.in_progress, url_record.status)
 
-        url_table.update(url_record.url, status=Status.done,
-                         increment_try_count=True, status_code=200)
+        url_table.check_in(url_record.url, Status.done,
+                           increment_try_count=True, status_code=200)
 
-        url_record = url_table[url_record.url]
+        url_record = url_table.get_one(url_record.url)
 
         self.assertEqual(200, url_record.status_code)
         self.assertEqual(Status.done, url_record.status)